+++ backend/src/main/java/com/kiwi/features/nodes/exceptions/NodeNotFoundException.java
@@ -3,7 +3,7 @@ package com.kiwi.features.nodes.exceptions;
 import jakarta.validation.constraints.NotNull;
 
 public class NodeNotFoundException extends RuntimeException {
-    public NodeNotFoundException(@NotNull int id) {
+    public NodeNotFoundException(@NotNull Long id) {
         super(String.format("Cannot found node with id %s", id));
     }
 }
